Enlarge the *Status columns - tinyint is not sufficient

+++ src/Dfc.CourseDirectory.Database/Pttcd/Tables/CourseRuns.sql
@@ -2,7 +2,7 @@
 (
 	[CourseRunId] UNIQUEIDENTIFIER NOT NULL CONSTRAINT [PK_CourseRuns] PRIMARY KEY,
 	[CourseId] UNIQUEIDENTIFIER NOT NULL CONSTRAINT [FK_CourseRuns_Course] FOREIGN KEY REFERENCES [Pttcd].[Courses] ([CourseId]),
-	[CourseRunStatus] TINYINT,
+	[CourseRunStatus] INT,
 	[CreatedOn] DATETIME,
 	[CreatedBy] NVARCHAR(MAX),
 	[UpdatedOn] DATETIME,
